+++ examples/live_language_change/server_app.R
@@ -1,3 +1,6 @@
+#' This script demonstrates how to use shiny.i18n Translator object
+#' to change the language dynamically on the server side.
+
 library(shiny)
 library(shiny.i18n)
 
@@ -12,7 +15,7 @@ server <- shinyServer(function(input, output, session) {
 
   i18n <- reactive({
     selected <- input$selected_language
-    if (length(selected) > 0 && selected %in% translator$languages) {
+    if (length(selected) > 0 && selected %in% translator$get_languages()) {
       translator$set_translation_language(selected)
     }
     translator
@@ -32,7 +35,7 @@ server <- shinyServer(function(input, output, session) {
         sidebarPanel(
           selectInput('selected_language',
                       i18n()$t("Change language"),
-                      choices = translator$languages,
+                      choices = translator$get_languages(),
                       selected = input$selected_language),
           sliderInput("bins",
                       "Number of bins:",